move shoelace out of PrescribeWorkflow
