+++ packages/elements/src/photon-multirx-form/components/PrescribeWorkflow.tsx
@@ -1,16 +1,36 @@
 import { gql } from 'graphql-tag';
-import { PhotonAuthorized } from '../photon-authorized';
-import type { FormError } from '../stores/form';
-import tailwind from '../tailwind.css?inline';
-import { checkHasPermission } from '../utils';
-import { AddPrescriptionCard } from './components/AddPrescriptionCard';
-import { DraftPrescriptionCard } from './components/DraftPrescriptionCard';
-import { OrderCard } from './components/OrderCard';
-import { PatientCard } from './components/PatientCard';
-import { PharmacyCard } from './components/PharmacyCard';
-import styles from './style.css?inline';
-import clearForm from './util/clearForm';
-import { formatPatientWeight } from './util/formatPatientWeight';
+import { PhotonAuthorized } from '../../photon-authorized';
+import type { FormError } from '../../stores/form';
+import { checkHasPermission } from '../../utils';
+import { AddPrescriptionCard } from './AddPrescriptionCard';
+import { DraftPrescriptionCard } from './DraftPrescriptionCard';
+import { OrderCard } from './OrderCard';
+import { PatientCard } from './PatientCard';
+import { PharmacyCard } from './PharmacyCard';
+import clearForm from '../util/clearForm';
+import { formatPatientWeight } from '../util/formatPatientWeight';
+import {
+  Alert,
+  AddressForm,
+  Button,
+  RecentOrders,
+  ScreeningAlertAcknowledgementDialog,
+  ScreeningAlertType,
+  SignatureAttestationModal,
+  Spinner,
+  TemplateOverrides,
+  Toaster,
+  triggerToast,
+  useDraftPrescriptions,
+  usePhoton,
+  usePrescribe,
+  useRecentOrders
+} from '@photonhealth/components';
+import { types } from '@photonhealth/sdk';
+import { Order, Prescription, PrescriptionState } from '@photonhealth/sdk/dist/types';
+import { GraphQLFormattedError } from 'graphql';
+import { createEffect, createMemo, createSignal, For, onMount, Ref, Show, untrack } from 'solid-js';
+
 const hasUsableAddress = (address?: {
   street1?: string;
   city?: string;
@@ -38,37 +58,6 @@ const fulfillmentNeedsAddress = (fulfillmentType?: string) => {
 const shouldBlockOrderWithoutAddress = (fulfillmentType?: string, hasAddress?: boolean) => {
   return fulfillmentNeedsAddress(fulfillmentType) && !hasAddress;
 };
-import {
-  Alert,
-  AddressForm,
-  Button,
-  RecentOrders,
-  ScreeningAlertAcknowledgementDialog,
-  ScreeningAlertType,
-  SignatureAttestationModal,
-  Spinner,
-  TemplateOverrides,
-  Toaster,
-  triggerToast,
-  useDraftPrescriptions,
-  usePhoton,
-  usePrescribe,
-  useRecentOrders
-} from '@photonhealth/components';
-import photonStyles from '@photonhealth/components/dist/style.css?inline';
-import { types } from '@photonhealth/sdk';
-import { Order, Prescription, PrescriptionState } from '@photonhealth/sdk/dist/types';
-import '@shoelace-style/shoelace/dist/components/alert/alert';
-import '@shoelace-style/shoelace/dist/components/icon-button/icon-button';
-import '@shoelace-style/shoelace/dist/components/icon/icon';
-import '@shoelace-style/shoelace/dist/components/switch/switch';
-import shoelaceDarkStyles from '@shoelace-style/shoelace/dist/themes/dark.css?inline';
-import shoelaceLightStyles from '@shoelace-style/shoelace/dist/themes/light.css?inline';
-import { setBasePath } from '@shoelace-style/shoelace/dist/utilities/base-path.js';
-import { GraphQLFormattedError } from 'graphql';
-import { createEffect, createMemo, createSignal, For, onMount, Ref, Show, untrack } from 'solid-js';
-
-setBasePath('https://cdn.jsdelivr.net/npm/@shoelace-style/shoelace@2.4.0/dist/');
 
 export type Address = {
   city: string;
@@ -597,12 +586,6 @@ export function PrescribeWorkflow(props: PrescribeProps) {
 
   return (
     <div ref={ref}>
-      <style>{tailwind}</style>
-      <style>{shoelaceDarkStyles}</style>
-      <style>{shoelaceLightStyles}</style>
-      <style>{styles}</style>
-      <style>{photonStyles}</style>
-
       <Show
         when={
           isScreeningAlertWarningOpen() && screeningAlerts().length > 0 && !overrideScreenAlerts()
